requirements.txt
aiofiles==24.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
duckduckgo_search==5.3.1b1
email_validator==2.1.1
etelemetry==0.3.1
executing==2.0.1
Faker==25.0.1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.13.1
Flask==3.0.1
flask-pytest==0.0.5
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
frontend==0.0.3
fsspec==2024.3.0
GeoAlchemy2==0.16.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-aiplatform==1.40.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gTTS==2.5.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
install==1.3.5
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.context==5.3.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.28
langchain-openai==0.2.14
langsmith==0.2.7
looseversion==1.3.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.3
openai==1.58.1
orjson==3.10.3
packaging==23.2
pandas==2.2.0
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pgvector==0.3.6
pillow==10.3.0
pluggy==1.4.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.2
prov==2.0.1
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
pydot==3.0.1
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==21.1.1
pytube==15.0.0
pytz==2023.3.post1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
redis==5.0.8
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scipy==1.14.0
setuptools==69.0.3
shapely==2.0.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
spotipy==2.23.0
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.37.2
Telethon==1.36.0
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.15.2
tqdm==4.66.1
traitlets==5.14.3
traits==6.3.2
transformers==4.38.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
whois==0.9.27
wolframalpha==5.0.1
xmltodict==0.13.0
